+++ include/physics/2d/Collision.h
@@ -6,7 +6,7 @@
 #define COLLISION_H
 
 #include <SFML/Graphics/Rect.hpp>
-#include "../physics/2d/CollisionAxis.h"
+#include "CollisionAxis.h"
 
 struct IncompleteCollision;
 class CollidableObject;
